Unified recursive traversal between transform types.
