+++ src/ForeverFactory/Generators/Transforms/Factories/FillWithEmptyStringTransformFactory.cs
@@ -1,48 +1,14 @@
 ﻿using System;
-using System.Linq;
 using System.Reflection;
 
 namespace ForeverFactory.Generators.Transforms.Factories
 {
-    internal class FillWithEmptyStringTransformFactory : ITranformFactory
+    internal class FillWithEmptyStringTransformFactory : BaseRecursiveTransformFactory
     {
-        public Transform<T> GetTransform<T>()
-            where T : class
-        {
-            var setMember = new Func<T, int, object>((instance, index) =>
-            {
-                FillPropertiesRecursively(instance, typeof(T));
-                return instance;
-            });
-            return new ReflectedFuncTransform<T>(setMember);
-        }
-
-        private static void FillPropertiesRecursively(object instance, IReflect type)
-        {
-            var propertyInfos = type.GetProperties(BindingFlags.Public | BindingFlags.Instance);
-            foreach (var propertyInfo in propertyInfos)
-            {
-                var buildFunction = GetBuildFunction(propertyInfo);
-                if (buildFunction == null)
-                    continue;
-
-                var propertyValue = buildFunction.Invoke();
-                propertyInfo.SetValue(instance, propertyValue);
-
-                FillPropertiesRecursively(propertyValue, propertyInfo.PropertyType);
-            }
-        }
-
-        private static Func<object> GetBuildFunction(PropertyInfo propertyInfo)
+        protected override Func<object> GetBuildFunctionForProperty(PropertyInfo propertyInfo, int index)
         {
             if (propertyInfo.PropertyType == typeof(string)) 
                 return () => string.Empty;
-
-            var parameterlessConstructor = propertyInfo.PropertyType
-                .GetConstructors()
-                .FirstOrDefault(x => x.GetParameters().Length == 0);
-            if (parameterlessConstructor != null) 
-                return () => parameterlessConstructor.Invoke(Array.Empty<object>());
 
             return null;
         }
